Work around a boost::wave #line management issue.
When a comment immediately follows a #line directive, the comment isn't
counted and the lines become mismatched. Add an empty newline with the
prefix headers to avoid this situation, compensating by subtracting one
from each line number for the main file.

Also updated glslang and SPIRV-Cross and suppress precision warning.