Fixes for adjusting SPIR-V bindings.
Force defaults to be written for flatbuffers when adjustable bindings are
used, otherwise the values can't be mutated if they start as the default
value. Fixed implementation of binding adjustment and added unit tests.

Updated glslang and SPIRV-Cross. Removed checks for binding indices when
unset since they are now being initialized to 0 for glslang.

Incremented release version.