Added varying block syntax.
Makes it simpler to declare outputs from one stage to inputs to another
stage, avoiding two declarations with filters. Use varying(output, input)
to declare the block (where both "output" and "input" are a stage name),
and every variable or block inside of the {} will be placed in the declared
stages with "out" or "in" prepended.

Additionally improved some error messages and updated glslang and
SPIRV-Cross. Incremented version number to 1.1.0.