Set buffer and texture uniform indices for Metal.
SPIRV-Cross uses the binding index for the uniform indices, so set them in
SPIR-V just before cross-compiling. Store this in the uniformIds array
that's already present.

Added module function to access the shader uniform IDs. This allows for
hooking up of uniforms in Metal. This also allows access in SPIR-V if
needed for whatever reason.