Added support for clip and cull distances.
Extract the size of gl_ClipDistance and gl_CullDistance builtin arrays and
store the values in the mslb file. This allows the client to determine if
the number of clip and cull distances are supported as well as handle any
state changes are required. (e.g. OpenGL) Also added #defines to determine
if the arrays are natively supported.

Incremented the version to 1.3.0.