Improved setting of version defines.
Now has the major version in the 100s digit and 10s and 1s digit for the
minor version, both for SPIR-V and Metal. This allows encoding of minor
versions >= 10, which can happen due to the . being a separator rather than
a decimal point.

SPIR-V version now passes the version it's actually compiling with. (1.4 as
of this writing)

Metal version is now parsed rather than using string comparisons. This
allows adoption of newer versions without modifying code.