Implemented parsing of fragment inputs.
Added new feature, FragmentInputs. If supported (Metal 2.0 for iOS or 2.3
for Mac) a "fragment" block can be declared to read fragment values. This
block behaves similarly to a uniform block for reading fragment values.
Additionally, the fragment_group render state can be declared for ordering
the fragment outputs. This is similar to subpasses and reading subpass
inputs in Vulkan, though it's different enough semanticly that it can't be
handled transparently.