Switched to use CMake targets for Boost dependencies.
This is more robust as it doesn't require manually importing dependencies
for static libraries, and avoids libraries like system that switched to
header-only. While boost_system has a stub library with the default build,
when building Boost through CMake (e.g. to create a universal binary on
Mac) the stub library isn't created.

CMake 3.5 is now required as that introduced the targets in FindBoost.

Incremented version to 1.5.4.