Switched to use GNUInstallDirs for install locations.
Should use more standard locations when installing across different Linux
systems and more configurability in general.

Added MSL_INSTALL_SET_RPATH CMake option, defaulting to ON, to set the
rpath on install. This ensures that a package built with shared libraries
can be relocated and the libraries will be found as expected.

Use a common install location for the CMake config files across all
platforms.