Added support for newer versions of SPIR-V.
The version config value is now respected for "spirv" shader type and and
used to control the version that is targeted when compiling. Other targets
use the latest version before cross-compiling.

Added support for new storage buffer representation when extracting
uniforms from SPIR-V, which has been in use since SPIR-V 1.6.

SPIR-V version define now follows the standard version code representation.
This also fixes the fact that the define improperly reported the latest
rather than the actual target, and the documentation didn't match how the
version was represented.

The version for Metal is now properly forwarded to SPIRV-Cross for
cross-compilation. This may fix some features not being enabled.

Updated glslang and SPIRV-Cross. Incremented the version to 1.8.0.